+++ src/chevah_compat/nt_filesystem.py
@@ -226,6 +226,22 @@ class NTFilesystem(PosixFilesystemBase):
         self._validateDrivePath(result)
         return six.text_type(result)
 
+    def isRoot(self, segments):
+        """
+        See `ILocalFilesystem`.
+        """
+        normalized_segments = self.getSegments(self.getPath(segments))
+        if self._lock_in_home:
+            return super().isRoot(normalized_segments)
+
+        if normalized_segments in [[], ['.'], ['..']]:
+            return True
+
+        return (
+            len(normalized_segments) == 1
+            and normalized_segments[0].lower() in self._allowed_drive_letters
+        )
+
     # Windows allows only 26 drive letters and is case insensitive.
     _allowed_drive_letters = [
         'a',
@@ -498,6 +514,10 @@ class NTFilesystem(PosixFilesystemBase):
         if not self.process_capabilities.symbolic_link:
             raise NotImplementedError('makeLink not implemented on this OS.')
 
+        self._rejectRoot(
+            link_segments, 'Creating a link as the root folder is not allowed.'
+        )
+
         target_path = self.getRealPathFromSegments(
             target_segments,
             include_virtual=False,
@@ -562,14 +582,14 @@ class NTFilesystem(PosixFilesystemBase):
         try:
             file_info = win32file.GetFileInformationByHandle(file_handle)
             (
-                attributes,
-                created_at,
-                accessed_at,
-                written_at,
+                _attributes,
+                _created_at,
+                _accessed_at,
+                _written_at,
                 volume_id,
-                file_high,
-                file_low,
-                n_links,
+                _file_high,
+                _file_low,
+                _n_links,
                 index_high,
                 index_low,
             ) = file_info
@@ -621,6 +641,10 @@ class NTFilesystem(PosixFilesystemBase):
         """
         See `ILocalFilesystem`.
         """
+        self._rejectRoot(
+            segments,
+            'Setting attributes for the nt root folder is not allowed.',
+        )
         with self._windowsToOSError(segments):
             if 'uid' in attributes or 'gid' in attributes:
                 raise OSError(errno.EPERM, 'Operation not supported')
@@ -808,6 +832,11 @@ class NTFilesystem(PosixFilesystemBase):
 
         For symbolic links we always force non-recursive behaviour.
         """
+        self._rejectRoot(
+            segments,
+            'Deleting the nt root folder is not allowed.',
+        )
+
         path = self.getRealPathFromSegments(segments, include_virtual=False)
         path_encoded = self.getEncodedPath(path)
         try:
@@ -854,6 +883,10 @@ class NTFilesystem(PosixFilesystemBase):
         """
         See `ILocalFilesystem`.
         """
+        self._rejectRoot(
+            segments,
+            'Setting owner for the nt root folder is not allowed.',
+        )
         path = self.getRealPathFromSegments(segments, include_virtual=False)
         encoded_path = self.getEncodedPath(path)
         try:
@@ -877,7 +910,7 @@ class NTFilesystem(PosixFilesystemBase):
                 )
                 d_acl = security_descriptor.GetSecurityDescriptorDacl()
 
-                user_sid, user_domain, user_type = (
+                user_sid, _user_domain, _user_type = (
                     win32security.LookupAccountName(None, owner)
                 )
                 flags = (
@@ -950,10 +983,14 @@ class NTFilesystem(PosixFilesystemBase):
         """
         See `ILocalFilesystem`.
         """
+        self._rejectRoot(
+            segments,
+            'Adding group for the nt root folder is not allowed.',
+        )
         path = self.getRealPathFromSegments(segments, include_virtual=False)
         encoded_path = self.getEncodedPath(path)
         try:
-            group_sid, group_domain, group_type = (
+            group_sid, _group_domain, _group_type = (
                 win32security.LookupAccountName(None, group)
             )
         except win32net.error:
@@ -988,10 +1025,14 @@ class NTFilesystem(PosixFilesystemBase):
         """
         See `ILocalFilesystem`.
         """
+        self._rejectRoot(
+            segments,
+            'Removing group for the nt root folder is not allowed.',
+        )
         path = self.getRealPathFromSegments(segments, include_virtual=False)
         encoded_path = self.getEncodedPath(path)
         try:
-            group_sid, group_domain, group_type = (
+            group_sid, _group_domain, _group_type = (
                 win32security.LookupAccountName(None, group)
             )
         except win32net.error:
@@ -1018,7 +1059,7 @@ class NTFilesystem(PosixFilesystemBase):
                 return None
             index_ace_to_remove = -1
             for index in range(ace_count):
-                ((ace_type, ace_flag), mask, sid) = dacl.GetAce(index)
+                ((_ace_type, _ace_flag), _mask, sid) = dacl.GetAce(index)
                 if group_sid == sid:
                     index_ace_to_remove = index
                     break
@@ -1045,7 +1086,7 @@ class NTFilesystem(PosixFilesystemBase):
         encoded_path = self.getEncodedPath(path)
 
         try:
-            group_sid, group_domain, group_type = (
+            group_sid, _group_domain, _group_type = (
                 win32security.LookupAccountName(None, group)
             )
         except win32net.error:
@@ -1066,7 +1107,7 @@ class NTFilesystem(PosixFilesystemBase):
                 # Nothing in the list.
                 return False
             for index in range(ace_count):
-                ((ace_type, ace_flag), mask, sid) = dacl.GetAce(index)
+                ((_ace_type, _ace_flag), _mask, sid) = dacl.GetAce(index)
                 if group_sid == sid:
                     return True
         return False